requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-pythonlibcore
openai
pytest